调整JsonResp中的error构造器，修改MemberController中的返回值以适配JsonResp，deleteMapping新增校验，修复MemberDAO中因日期空校验引起的异常错误，修复MemberDAO中相似字段无法匹配问题

+++ src/main/java/org/example/demo05/controller/MemberController.java
@@ -11,7 +11,6 @@ import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
-import java.util.Map;
 
 //@CrossOrigin(origins = "http://localhost:3000", allowCredentials = "true")
 @RestController
@@ -32,22 +31,40 @@ public class MemberController {
             PageInfo<?> pageInfo = new PageInfo<>(members);
             return JsonResp.success(pageInfo);
         } catch (Exception e) {
-            return JsonResp.error(e.toString());
+            return JsonResp.error(500, e.toString());
         }
     }
 
     @PostMapping
-    public Map<Boolean, String> addMember(@RequestBody Member member) {
-        return memberService.addMember(member);
+    public JsonResp addMember(@RequestBody Member member) {
+        try {
+            int resp = memberService.addMember(member);
+            return JsonResp.success(resp);
+        } catch (Exception e) {
+            return JsonResp.error(500, e.toString());
+        }
     }
 
     @PutMapping
-    public Map<Boolean, String> updateMember(@RequestBody Member member) {
-        return memberService.updateMember(member);
+    public JsonResp updateMember(@RequestBody Member member) {
+        try {
+            int resp = memberService.updateMember(member);
+            return JsonResp.success(resp);
+        } catch (Exception e) {
+            return JsonResp.error(500, e.toString());
+        }
     }
 
     @DeleteMapping
-    public Integer deleteMember(Integer[] ids) {
-        return memberService.deleteMember(ids);
+    public JsonResp deleteMember(Integer[] ids) {
+        if (ids != null && ids.length == 0) {
+            return JsonResp.error(400, "id为空");
+        }
+        try {
+            int res = memberService.deleteMember(ids);
+            return JsonResp.success(res);
+        } catch (Exception e) {
+            return JsonResp.error(500, e.toString());
+        }
     }
 }
